fix logfile

+++ boinconpi.sh
@@ -186,10 +186,11 @@ echo "Updating APT..."
 apt -y update >> $LOGFILE 2>&1
 echo "Cleaning up any previous BOINC installation attempts"
 apt -y --purge boinc boinc-client boinc-client-opencl boinc-clicnt-nvidia boinctui >> $LOGFILE 2>&1
-rm -r /var/lib/boinc-client >> $LOGFILE 2>&1
+[ -d "/var/lib/boinc-client" ] && rm -r /var/lib/boinc-client >> $LOGFILE 2>&1
 echo "Installing BOINC.."
-apt -y install boinc boinc-client-opencl boinc-client-nvidia >> $LOGFILE 2>&1
-apt -y install boinctui >> $LOGFILE 2>&1
+for i in boinc boinc-client-opencl boinc-client-nvidia boinctui; do
+  sudo apt-get install -y $i >> $LOGFILE 2>&1
+done
 echo "Contacting BOINC servers.."
 #attach to account manager
 boinccmd --acct_mgr attach "$BAMURL" "$USERNAME" "$PASSWORD" >> "$LOGFILE" 2>&1
